rsets: add castPosition helper function and update
